+++ lib/components/animation/bubble_button.dart
@@ -32,10 +32,12 @@ class CircleColor {
   final Color end;
 
   @override
-  bool operator ==(dynamic other) {
-    if (other.runtimeType != runtimeType) return false;
-    return other is CircleColor && start == other.start && end == other.end;
-  }
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other.runtimeType == runtimeType &&
+          other is CircleColor &&
+          start == other.start &&
+          end == other.end;
 
   @override
   int get hashCode => Object.hash(start, end);